Refactor value objects to follow ActiveModel conventions (#136)
- Add Flex::ValueObject
- Refactor the following value objects to subclass this one:
  - Address
  - Money
  - Name
  - ValueRange
  - YearQuarter

Models that subclasses primitive types instead of ValueObject:
- TaxId (subclasses String)
- USDate (subclasses Date)

In the process of refactoring ValueRange, added a feature where
flex_attribute works with built in rails attribute types as well.

## Context

I don't know why I went down this rabbit hole. I think I was trying to
implement LeavePeriod and for some reason I thought it'd be better to
get this over with first.

There's one test that I couldn't get to pass, so I had to comment that
out and created a ticket for it
https://linear.app/nava-platform/issue/TSS-175/make-yearquarter-more-strict-about-types-rather-than-liberally-casting