Remove path_func option from case index components (#215)
- Remove path_func option from case index components
- Move all passport-related files out of strata and under spec/dummy/app

## Context

As I was implementing the case row component in OSCER in
https://github.com/navapbc/community-engagement-medicaid/pull/65 I was
finding that it's awkward to require users of the SDK to need to know
about the path_func option given that it's only used by Lookbook
previews, and the only reason we used it was because in the SDK we
couldn't guarantee that a concrete subclass of Case exists that has
routes defined for it (so that the polymorphic_path calls in the view
component wouldn't error). This PR removes the path_func option and
instead adds some magic logic to the index_preview_component to search
for a concrete subclass of Case that can have polymorphic_path called on
it. I think the magic code is not great but it's worth the tradeoff to
keep the public SDK API cleaner.

While working on this I noticed a bunch of passport-related files were
in the wrong place. They shouldn't be in the Engine folder since they
are unrelated to the SDK. Instead they should be in the spec/dummy/app
folder.