Change Determination to store array of reasons (#240)
## Changes

A determination will now store an array of reasons.

## Context

While working with determinations, a person could have many reasons that
factor into a determination. A determination is also immutable, and the
intention is for a determination to supersede a previous one on the
aggregate root. This was discussed with team members on October 31st,
2025.